Test trials to check times for running experiments
Kmeans takes for all images:
35 seconds for 3 clusters
70 seconds for 5 clusters
185 seconds for 10 clusters

Also, the martin index evaluation code does not work for for spectral or FCM algorithms as it gives errors. Also i'm a bit concern that is giving such good values for kmean (0.9 and above) for all images. So we might want to check that? once everthing is solved, someone needs to spend a couple of hours to run the code for all and produce csv files for all martin indexes for images with all combinations of cluster numbers/groundtruths